Big overhaul.
This version adds support for editions which required changes to most files. Lots of other changes as well such as adding rudimentary debugging and testing support in bib_import.py

Known issues: adding and removing books for collections doesn't work. Jordan's Wheel of Time has an issue with series numbering (this could be an old bug).
: